Mark call for evidence document type non-indexable
This is required because Whitehall calls the search API directly, so we do not want the publishing API to also attempt to index calls for evidence.
This is technical debt and we want to remove this behaviour from Whitehall in the future, at which time we will be able to migrate to the new indexing process.