Add OIM Project and SAU Referral case types
The Competition and Markets Authority have recently added Office for the Internal Market projects and Subsidy Advice Unit referrals as new case types to their CMA case specialist document type. This adds the new case types to the search API, and also adds a "Not Appliocable" market sector because the market sector field does not apply to the new case types.