Provide a rake task for updating an Elasticsearch index schema in place when Search API configuration has changed
This rake task has been added because at the moment even purely additive changes to elasticsearch schema configuration require a full reindex. Full reindexes require locking the index for significant periods of time so have to be run out of hours for production and can block integration for long periods. This is causing pain when developing applications that consume the Search API.

The task is idempotent and Elasticsearch will not permit changes to the index that would attempt to change the type of already existing data. Any such changes result in a “Bad Request” error from Elasticsearch and these will be reported to the rake operator. In such cases a full re-index would be required.