Remove require from test as it's hiding other require errors
Having require 'active_support/time' in this spec means this
was available when running the entire suite. This covered
a require error in production code that made it into prod.

The test for that code only failed when run in isolation and the require from
this test hadn't happened.

Current state of this commit is the tests are failing - as they should be.