Add GA pipeline components
As part of the migration to GA4 into search-api. I have created an orchestrator class RelevanceReportGenerator which is responsible for fetching the data from GA, consolidating that data and then converting this into an elasticsearch format ready for ingestion via another cron job.

The flow is we make the call to GA4, we set the metrics and dimensions we want from GA, we obtain this data, we analyse this data and then filter out the records we don't want (page not found pages and smart pages and non relative path pages), we then finally serialise this into an elastic search format. This is then sent to ElasticSearch ready for another cron job to take over to update the search-api popularity page.

The code was ported from https://github.com/alphagov/search-analytics-ga4 and changes have been made

The reason for porting the code into here is because it’s easier to run the task if part of search-api - can be done on a pod in K8s, whereas running search-analytics from GitHub actions was disastrous for catching failures, it’ll be easier to maintain as a single codebase and not a separate thing for teams to look at