Update page_traffic rake task
Prior to this change the process was to grab the GA file from S3, however now we do all of this logic within search-api, we dont need the S3 bucket. So now we just send the output from the RelevanceReportGenerator straight to the PageTrafficLoader, in the page_traffic rake task. This is triggered via a cron job at around 5am every day.