Update updater to not pass entire document into sidekiq jobs
Search API is using up all the Redis memory in integration.
The main culprit is the PopularityWorker - this serialises a large batch of data
to update, it includes the entire `_source` json - this make the payload very large.

It is also completely unnecessary as the item could just as easily be pulled from
elastic.

This change updates the updater to only pass the documents identifier to the
updaters that inherit from it.

It then updated the popularity worker to lookup the document in the index as part of the
job before updating the popularity.

The work to update the supertype worker will come in the following commit.