Add new endpoint to query specialist-finder index
New parallel stack to the existing search endpoint so we don't interfere with existing search functionality.
When doing the testing, we've noticed that certain things are configured only for government_test index but not for the govuk (and hence also our new index) - such as spelling typo suggestions, and organisation expansion.

Should the tests be more explicit about calling this out?