Refactor search server initialisation and remove dead code
The search server required a number of different index name arguments to its initialiser, including a specific argument for the govuk index name.

This was in large part due to an unused method on the search server class that required a subset of the names.

Removing that method allows us to simply pass all of the available index names to the search server in a single argument.

This will make it easier to potentially add new indexes to Elasticsearch in the future.