Remove deprecation marker on expanded search result field configuration and explain usage
Given that we don't know why this behaviour was marked deprecated, and downstream apps depend on it, it seems reasonable to remove the deprecation notice.

This commit also adds a usage example from Finder Frontend to make it clear what purposes the configuration can be used for. Arguably finder frontend could look up the label from the finder content item instead of requiring the label to be part of the search result, but that is unlikely to be a simple change.