Clear Sidekiq::Jobs after using Sidekiq::Testing.fake!
This change has been done to resolve a flaky test [1] where we were
seeing an error related to expectations of the existence of a Sidekiq
hash:

```
     Failure/Error: work["queue"] == self::QUEUE_NAME && work["payload"]["class"] == to_s

     NoMethodError:
       undefined method `[]' for nil:NilClass
     # ./lib/base_job.rb:30:in `block in active_jo
```

This was caused by the test that is the subject of this commit leaving
state in Sidekiq. This is caused by Sidekiq::Testing.fake! which will
leave information in the Sidekiq::Job.jobs array. Running
Sidekiq::Job.clear_all in a before block of a single test file will only
clear out the state before the tests run but not do anything to resolve
state left behind by the tests. It looks like this has
revealed a bug inadvertently, which we can fix in a separate commit.

As this is the only place we use Sidekiq::Testing.fake! we should only
need to clear jobs after this.

[1]: https://github.com/alphagov/search-api/actions/runs/11909087045/job/33185768393