Update Sidekiq API usage and add tests for PageTrafficJob
These methods previously had no tests and we discovered they were
recently broken. Example error:

```
NoMethodError: undefined method `jobs' for Sidekiq::Job:Module (NoMethodError)
    Sidekiq::Job.jobs.any? do |_, _, work|
                ^^^^^
/app/lib/base_job.rb:29:in `active_jobs?'
/app/lib/base_job.rb:24:in `block in wait_until_processed'
/app/lib/base_job.rb:19:in `wait_until_processed'
```

When investigating this it transpired that the `jobs` method is only
available in Sidekiq::Testing [1] so this code never worked in practice.

I've now updated the code to use aspects of Sidekiq::API code. It is
worth noting that Sidekiq strongly suggest that you should not code from
Sidekiq::API in your business logic [2] unfortunately that ship had
already sailed with the `wait_until_processed` method - and is very much
a bad smell of this part of the codebase that should really be
reconsidered.

Because of this unexpected Sidekiq::API usage, the Sidekig::Testing API
isn't available so there was the option to either test directly through
Redis or do mocking. I went for mocking. This does mean that the tests
I've added are rather contrived, and having to hope that what I've set
up will be enough to catch any changes in Sidekiq::API.

Of course, there are still methods that lack unit tests for the class
after this, I just added the ones for the static methods that had fallen
victim to causing failures.

[1]: https://github.com/sidekiq/sidekiq/blob/0e9419a7fe5a543794d896246e0eac594ea19235/lib/sidekiq/testing.rb#L300-L302
[2]: https://github.com/sidekiq/sidekiq/blob/0e9419a7fe5a543794d896246e0eac594ea19235/lib/sidekiq/api.rb#L10-L19