Require "sentry-sidekiq" gem
Previously this gem was installed to resolve the error of:

```
Warning: GovukError is not configured to track Sidekiq errors, install the sentry-sidekiq gem to track them.
```

in [1].

However this wasn't sufficient because Search API doesn't automatically
require gems, so while the gem was available it wasn't actually used.

By requiring it we can resolve this error.

[1]: https://github.com/alphagov/search-api/pull/2494