Add migrated app configuration to Search API and always include results from migrated apps
This allows us to safely migrate document formats to the govuk index in Whitehall. With the current configuration that isn't possible because the format migrator doesn't respect publishing app configuration within the migrated section of the migrated_formats.yaml file.

We have considered maintaining the apparent original intention of the file which was to migrate items on a format by format basis, but applying the format migration for specific publishing apps added substnatial complexity to the Elasticsearch filter which might have affected search performance.

Because we can now just declare an app migrated, we no longer need the configuration for restricting indexable formats based on their publishing app either, so that code has been removed.