Mark all publication types as migrated to the GOVUK index
I added a couple of comments to warn future devs against removing some migrated formats that are not actually used in the govuk index.

Those formats don’t exist on the govuk index because they’ve been renamed to something else, or split into sub-types, during the migration from the government index. It therefore feels weird having them in the migrated list. However, if we take them out, we’ll get duplicate results in searches, because Search API will happily serve the new version of the record from the govuk index with one format and the old record from the govuk index in the old format in the same result set.