Reset Sidekiq concurrency limits to sensible defaults
We are going to revert the helm config to the original two replicas running Sidekiq, with 12 threads each. We now know we can safely use them all without overburdening Elasticsearch, so this offers 12 threads to the default queue and 12 to the bulk queue. There is an argument that we don't really need the limits at all, but it is probably worthwhile to reserve some threads for each queue. We should re-evaluate this choice at a later date.

One thing that helped speed up indexing was to reduce Elasticsearch's refresh interval to once every 30 seconds. Refreshing is the process by which Elasticsearch makes indexed data available for searching and it is moderately resource intensive. It doesn't make sense to run this every second and reducing the frequency helps to improve indexing throughput.