Include document collection slugs in govuk documents
This field was present on the government index, but was dropped during the migration of Whitehall content to the govuk index because it was not used in any of the finders.

However, a team from the MoJ depended on the field for one of their applications, which consumes Search API, so we're reinstating it.