Retry processing on search_api_to_be_indexed queue
The `Indexer::MesageProcessor` class handles processing RabbitMQ
messages specifically pulled from the `search_api_to_be_indexed` queue.

This commit reworks the class so that it retries message processing up
to 5 times if an error is raised.

Previously the class handled two errors:

* If a `ProcessingError` is raised, we simply report the exception and
  stop trying to process. However looking in the codebase, it seems like
  this `ProcessingError` is defined but never actually raised anywhere.
  So this rescue block wouldn't be executed
* If a `StandardError` is raised, we sleep for 1 second and retry the
  message indefinitely. There's no logic to break out of this retrying
  loop

This change just rescues from `StandardError` and will inspect the
message to see how many times it's been retried already. If it's <= 5,
we retry, otherwise we report the error and ack the message which puts
it on a dead letter queue and won't be processed again.

Note that when we call `message.retry`, there's some RabbitMQ
configuration already in place which sets a delay of 60s between
retries. See the government-infrastructure repo for more information
about this but essentially it works like so:

1. A message is processed on the `search_api_to_be_indexed` queue. An error
   is raised and we call `message.retry` (which actually calls `message.discard`)
2. This moves the message to the configured dead-letter exchange where
   it sits on a queue and has a TTL of 60 seconds. Nothing listens on
   this queue so it's never processed
3. After 60 seconds, it's automatically routed back to the
   `search_api_to_be_indexed` queue for re-processing