Extend AmendJob class to optionally reschedule on failure
This job class is used in multiple places across the codebase to enqueue
a Sidekiq job to amend a document in the index.

We want to remove the Sidekiq dependency when processing jobs from
RabbitMQ, which is one of the places this job is enqueued from.

The way to do this with the least amount of friction is to keep the job
class, but allow it to be configured to either retry the Sidekiq job or
re-raise the error when caught.

This means that for parts of the app that still want to use Sidekiq
don't need any changes made, and continue to be able to enqueue and
retry the job on Sidekiq, but for the RabbitMQ consumer we can pass in
`reschedule_on_failure: true` and let the exception bubble up to where
we can catch it and retry the RabbitMQ message.