Disable Sidekiq processing on RabbitMQ message consumer
In order to remove the Sidekiq dependency and move retrying logic to
RabbitMQ, we need ot change this `ChangeNotificationProcessor` in a
couple of ways.

Firstly we call the `perform` method directly, rather than calling
`perform_async` which will enqueue the job in Sidekiq. Calling the
method directly will do the work synchronously, but means we can still
call `perform_async` elsewhere in the codebase.

Passing in `reschedule_on_failure: false` will re-raise an exception in
the `AmendJob` class if one is raised. Omitting this argument will cause
another Sidekiq job to be enqueued if an error is raised, which
obviously don't want if we're removing this Sidekiq dependency.

So instead we let the exception re-raise and the queue consumer code
(`Indexer::MessageProcessor`) will rescue it and handle it
appropriately.