Remove 'how search works' doc
This was very light on detail and its title intefers with the new
canonical document we're setting up in the Developer Docs.

One useful thing it did hint at was the 'batch_search' endpoint
which is otherwise undocumented. That has now been added to the
"using the search API" doc.