Create new retry/discarded queues in dev rake task
In the same way that we do automatic retries on RabbitMQ for messages on
the `search_api_to_be_indexed` queue, we want to apply this to the two
other queues we listen on:

* `search_api_bulk_reindex`
* `search_api_govuk_index`

So this commit updates the `message_queue:create_queues` rake task to do
that. In order to cut down on a lot of repetition - because we need to
create two new exchange and three new queues for each existing queue -
we can all the heavy-lifting in a loop.