Remove Sidekiq from PublishingEventJob
This Sidekiq job currently only exists for messages to be pulled off
RabbitMQ and then immediately enqueued on Sidekiq for processing.

We're going to remove the entire Sidekiq aspect of this and instead
process the message directly in the queue consumer.

So this removes all Sidekiq-related code from the job and instead the
consumer will instantiate the class directly and call `perform` on it.

A side-effect of this is that we no longer need to partition the
incoming RabbitMQ messages. This logic only existed so that messages
matching a certain routing key would be put on a specific Sidekiq queue,
whilst the others were put on the default queue. All other processing
logic is the same.