Implement retries in PublishingEventProcessor
This is a continuation of the work to remove Sidekiq from the consumer
classes that process jobs coming off the RabbitMQ queues.

Currently we pull jobs off and immediately enqueue them on Sidekiq,
which can put a lot of strain on Redis if we're processing a lot of
items at once (e.g. in a bulk reindex of GOV.UK content).

So instead we've used the concurrency features offered by
`govuk_message_queue_consumer` and we'll process messages directly in
the consumer. This has the downside that we need to implement retry
logic ourselves in the consumer class, which is what this commit does.

We'll retry a message up to 5 times if an exception is raised during
processing. We catch the exception and call `message.discard` which will
queue it for retrying after a short time (this is configured in
RabbitMQ).

One other significant change in this commit is that we've removed the
logic to pull multiple messages off a queue at once.

Previously the consumer had a `process` method which took an argument of
one or more messages, and would loop through those messages adding them
to a bulk update in ElasticSearch, and finally committing that update.

However this makes implementing retries a bit tricky, because if
anything went wrong with the bulk update, we'd need to retry each
message individually (by calling `discard` on each one). Doing this
would requeue each message on RabbitMQ where they'd then be processed
individually when it comes time to retry them. So we'd lose any
advantage of the bulk updating if a single exception was raised.

Furthermore - and more significantly - there's no evidence anywhere to
show that messages ever get enqueued on RabbitMQ in groups, or even that
it's possible to pull multiple messages off a queue simultaneously. The
code in this consumer which handles multiple messages was written around
8 years ago[1], so perhaps it was a feature then but isn't any longer.
There used to be the concept of a batch consumer in
`govuk_message_queue_consumer` but that was a separate consumer class,
and support for it was completely removed in 2022 anyway[2].

So it feels like it makes sense to refactor this
`PublishingEventProcessor` consumer class to handle just one message at
a time.

[1] https://github.com/alphagov/search-api/commit/94aa4ce2831e57fa023900cc92aefcfc795163e6
[2] https://github.com/alphagov/govuk_message_queue_consumer/pull/73