Move error classes out of PublishingEventMessageHandler file
These error classes were dumped in this file so that they were
namespaced to `GovukIndex` (i.e. outside of the
`PublishingEventMessageHandler` class).

This isn't really the best place for them, so they've been moved.

Some were moved to specific classes where they were used, but the more
generic classes were moved to a `govuk_index` file and included in the
`GovukIndex` namespace there.