Don't try to process messages that'll never be processed
There are certain messages that can come off RabbitMQ which we'll never
be able to process:

* A content type that is _not_ `external_content` where no `base_path` is set
* An `external_content` content type with no `details.url` set

These will just go through processing until an exception is raised, in
which case we'll retry the messages 5 times until we give up.

Rather than doing all this, we can just return early in the processing
and call `message.done` while logging a message.

In doing this, we're able to remove the `valid!` check in the
presenters. These validations are no longer needed as messages will have
been rejected prior to this code being executed.

A few of the tests needed fixing here because they didn't have a
`base_path` attribute set. The ES index works a little bit unexpectedly
in this app in that a document's `id` field is either a `content_id` or
a `base_path`. For the specs that needed updating, I switched from using
the `content_id` to `base_path` for the ID. This allows the content
items to be processed by the queue consumer (because the `base_path` is
now set), and allows those documents to be indexed and queried in ES by
that `base_path` instead of the `content_id`.