Index full attachment URL for different base paths
Previously, the search API assumed that the URL for a document
part could be constructed by appending the part's 'slug' to the
parent document's base path.

This assumption is incorrect for some formats. For example, with
statistics bulletins, the HTML attachment may be located under
`/government/publications/...` while the parent page is under
`/government/statistics/...`. This discrepancy would cause
search results to link to broken pages for these formats, and so
these attachments with differing base paths weren't indexed at
all.

This commit addresses removes the check that avoided indexing
in these situations, as we now store the full URL of an HTML
attachment in the `link` attribute of the corresponding part.