Add support for organisations and taxons to finders
If a finder is tagged to an organisation, the organisation will appear
in a metadata box at the top of the page (e.g. "From: HM Revenue &
Customs").

If a finder is tagged to a taxon, it will have breadcrumbs at the top of
the page (e.g. Home > Money > Dealing with HMRC > Tax agent and adviser
guidance)

This adds the ability for finders to configure organisations and taxons,
in the same way they already can for parents and related items.