Migrate the people finder to search-api
Migrating from Whitehall:
https://github.com/alphagov/whitehall/blob/6372d382990abd4d42e8696cf50204005382d5ca/lib/finders/people.json

As per the Whitehall docs:
https://github.com/alphagov/whitehall/blob/75f4d532fa29443bd316fd395afb471a3d00c2e7/docs/finders.md#how-to-publish-a-finder-in-whitehall

Non-specialist finders are currently spread across Whitehall and Search API. A decision was made a while ago for these finders to live in Search API only, but the migration never happened. We're now migrating these finders and will delete their configurations from Whitehall.

We've translated the JSON configs to their YAML equivalents as best as possible. There is no email signup page for this finder. The only rendering difference should be in the sorting, where by default we were sorting by public_timestamp in the earlier version of this finder, but taking inspiration from the statistics finder in Search API, there's no harm in offering to sort by both ASC and DESC order.
There are no filters/facets for this finder.

Jira: WHIT-1833