Add a 'claim_route' rake task
These finder routes are owned by Whitehall and so trying to
publish the finders using the rake task in search API would lead
to a UnprocessableEntity error.

We've therefore added a rake task we can run first, to claim
ownership of a route in search API _before_ attempting to publish
content to that route.

We'll be able to delete this rake task after we've run the
migration.