Add Indexer tests
It's not possible to nicely test for other expectations when an error is
raised. You need to wrap any other expectations inside the expect for
the raised error. Otherwise the code your trying to test either isn't
reached or isn't able to report back.

These two posts explain it better:
[stackoverflow](https://stackoverflow.com/questions/16858899/how-do-i-expect-something-which-raises-exception-in-rspec)
[github issue](https://github.com/rubocop/rubocop-rspec/issues/379)