Add tests for AutocompletePresenter
These tests attempt to imagine what that data structure of es_response
should be based on the code in the `suggestions` method.

`ignored_metadata` has been added to the autocomplete array in the test
data because the code looks for `result[1]` to get the `options` which
implies that there is something unused by this class in `result[0]`.