Remove NotIdentifiable error
This is a custom error used inside search-api.
It used to be raised when base paths were
missing from the publishing event payload,
but since this is an optional field, these
cases are now handled more gracefully. Now
these custom errors are no longer used and
can be removed.

PR that introduced more graceful handling
of empty base paths is here:
https://github.com/alphagov/search-api/pull/3237